Add files via upload
Hello, and welcome to my second data analysis project in Excel. This analysis was performed using just Excel and makes use of open adidas sales data.

What was the aim of this project?
To monitor key performance indicators of sales for the business over the specified period of time the data covers.

This analysis answers questions like:

1.	How many units are sold on a monthly basis?
2.	What is the average operating revenue?
3.	What are the revenues generated from sales to key customers, and what products contribute the most to sales across each branch of the business?’

With these insights, a business would be able to make more data driven decisions.